usuario_mejorado_api

+++ routes/api/usuarios.js
@@ -2,6 +2,9 @@ var express = require('express');
 var router = express.Router();
 const bcrypt = require("bcrypt");
 const Usuario = require('../../database/models/usuario');
+//const users = require('../database/users');
+const USUARIOSCHEMA = Usuario.schema;
+var valid = require("../../utils/valid");
 /* GET Usuarioe. */
 router.get('/', function (req, res, next) {
 
@@ -20,18 +23,40 @@ router.get('/', function (req, res, next) {
     })
 
 });
+router.post('/',async(req,res) => {
+   var params = req.body;
+   params["registerdate"]= new Date();
+  if(!valid.checkParams(USUARIOSCHEMA, params)){
+    res.status(300).json({
+      msn:"parametros incorrectos"
+    });
+    return;
+  }
+  if(!valid.checkEmail(params.email)) {
+    res.status(300).json({
+    msn:"Email invalido"
+  });
+  return;
+}
+if(!valid.checkPassword(params.password)) {
+  res.status(300).json({
+  msn:"password invalido"
+});
+return;
+}
+  var usuario = new   USUARIO(params);
+  var result = await usuario.save();
+  res.status(200).json(result)
+});
 
-router.post('/', function (req, res, next) {
 
+/*router.post('/', function (req, res, next) {
 
     const datos = {
         name: req.body.name,
         lastname: req.body.lastname,
         email: req.body.email,
-        //password: req.body.password,
-        //telefono: req.body.telefono,
-        //avatar: req.body.avatar,
-        //tipo: req.body.tipo,
+
     };
     bcrypt.hash(req.body.password, 10, (err, hash) => {
         if (err) {
@@ -56,10 +81,23 @@ router.post('/', function (req, res, next) {
         }
     });
 
-});
+});*/
+ /*router.patch('/',(req,res) => {
+   if (req.query.id == null) {
+     res.status(300).json({
+       msn:"error no existe id"
+     });
+     return;
+   }
+   var id = req.query.id;
+   var params = req.body;
+   Usuario.findByIdAndUpdate({_id: id}, params, (err,docs) => {
+     res.status(200).json(docs);
+   });
+});*/
 
-router.patch('/:id', function (req, res, next) {
-    let idUsuario = req.params.id;
+router.patch('/', function (req, res, next) {
+    let idUsuario = req.query.id;
     const datos = {};
 
     Object.keys(req.body).forEach((key) => {
@@ -78,7 +116,17 @@ router.patch('/:id', function (req, res, next) {
         });
 });
 
-router.delete('/:id', function (req, res, next) {
+router.delete("/",async(req,res) => {
+//  var id = req.query.id;
+  if(req.query.id == null){
+    res.status(300).json({
+      msn:"Error no existe el id"
+    });
+    return;
+  }
+  var r = await Usuario.remove({_id: req.query.id});
+  res.status(300).json(r);
+/*router.delete('/', function (req, res, next) {
     let idUsuario = req.params.id;
 
     Usuario.findByIdAndRemove(idUsuario).exec()
@@ -92,7 +140,7 @@ router.delete('/:id', function (req, res, next) {
             });
         });
 
-
+*/
 });
